pyproject.toml
[project]
name = "bootcode"
version = "0.1.0"
description = "Simple AI agent built from Boot.dev tutotial"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
